make sure float64 passed to redtoreg

+++ src/pygrib/_pygrib.pyx
@@ -220,7 +220,7 @@ def redtoreg(redgrid_data, lonsperlat, missval=None):
     If any values equal to specified missing value (``missval``, default NaN), a masked array is returned."""
     if missval is None:
         missval = np.nan
-    datarr = _redtoreg(lonsperlat.max(),lonsperlat,redgrid_data,missval)
+    datarr = _redtoreg(lonsperlat.max(),lonsperlat,redgrid_data.astype(np.float64),missval)
     if np.count_nonzero(datarr==missval):
         datarr = ma.masked_values(datarr, missval)
     return datarr
